refactor: move filenames from gateway to transport or protocol. Rename factory class to AgntcyFactory
