+++ tests/server/__server__.py
@@ -29,9 +29,9 @@ import asyncio
 import argparse
 from uvicorn import Config, Server
 from agntcy_app_sdk.factory import TransportTypes
-from agntcy_app_sdk.factory import GatewayFactory
+from agntcy_app_sdk.factory import AgntcyFactory
 
-factory = GatewayFactory()
+factory = AgntcyFactory()
 
 
 async def main(transport_type: str, endpoint: str, block: bool = True):
